pyproject.toml
[project]
name = "musy"
version = "0.1.0"
description = "Telegram inline bot for Yandex Music"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.4",
    "yandex-music>=3.0",
    "python-dotenv>=1.0",
    "aiohttp-socks>=0.9",
]

[project.scripts]
musy = "musy.__main__:cli"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["musy*"]
